Tests for combined workflow, iap and flow runner
