+++ dipy/workflows/tests/test_combined_workflow.py
@@ -1,38 +1,12 @@
 import numpy as np
-import os
 from nose.tools import assert_raises
 from os import path
 
 from nibabel.tmpdirs import TemporaryDirectory
 
 from dipy.workflows.combined_workflow import CombinedWorkflow
-from dipy.workflows.workflow import Workflow
-
-
-class DummyWorkflow1(Workflow):
-    def run(self, inputs, param1=1, out_dir='', output_1='out1.txt'):
-        return param1
-
-
-class DummyWorkflow2(Workflow):
-    def run(self, inputs, param2=2, out_dir='', output_1='out2.txt'):
-        return param2
-
-
-class DummyCombinedWorkflow(CombinedWorkflow):
-    def _get_sub_flows(self):
-        return [DummyWorkflow1, DummyWorkflow2]
-
-    def run(self, inputs, param_combined=3, out_dir='',
-            output_combined='out_combined.txt'):
-
-        dwf1 = DummyWorkflow1()
-        param1 = self.run_sub_flow(dwf1, inputs)
-
-        dwf2 = DummyWorkflow2()
-        param2 = self.run_sub_flow(dwf2, inputs)
-
-        return param1, param2
+from dipy.workflows.tests.workflow_tests_utils import DummyCombinedWorkflow,\
+    DummyWorkflow1, DummyWorkflow2
 
 
 def test_get_sub_flows():
@@ -59,7 +33,7 @@ def test_combined_run():
         inp = path.join(tmpdir, 'test.txt')
         np.savetxt(inp, np.arange(10))
 
-        ret_param1, ret_param2 = dcwf.run(inp)
+        ret_param1, ret_param2, _ = dcwf.run(inp)
         assert ret_param1 == param1_value
         assert ret_param2 == param2_value
